+++ Sources/CorePayments/Networking/HTTPResponseParser.swift
@@ -69,3 +69,26 @@ public class HTTPResponseParser {
         }
     }
 }
+
+extension HTTPResponseParser {
+
+    // to extract correlationId from GraphQL response
+    @_documentation(visibility: private)
+    public func parseGraphQLDictionary(_ httpResponse: HTTPResponse) throws -> [String: Any] {
+
+        guard httpResponse.status == 200 else {
+            throw NetworkingError.urlSessionError
+        }
+
+        guard let body = httpResponse.body else {
+            throw NetworkingError.noResponseDataError
+        }
+
+        let jsonObject = try JSONSerialization.jsonObject(with: body, options: [])
+        guard let topLevel = jsonObject as? [String: Any] else {
+            throw NetworkingError.invalidURLResponseError
+        }
+
+        return topLevel
+    }
+}