+++ Sources/CorePayments/PayPalCoreConstants.swift
@@ -6,7 +6,7 @@ public enum PayPalCoreConstants {
     // TODO: - Update release script to update this version #
     /// This property is exposed for internal PayPal use only. Do not use.
     /// It is not covered by Semantic Versioning and may change or be removed at any time.
-    public static let payPalSDKVersion: String = "2.0.0"
+    public static let payPalSDKVersion: String = "2.0.1"
     
     public static let callbackURLScheme: String = "sdk.ios.paypal"
 }