+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   * PayPalButtons
     * Remove `PaymentButtonSize` enum
     * Remove `size` property from `PaymentButton` base class
+    * Remove `PaymentButtonColor.GOLD` enum case
+    * Remove `PaymentButtonColor.SILVER` enum case
+    * Remove `PayPalCreditButtonColor.GOLD` enum case
+    * Remove `PayPalCreditButtonColor.DARK_BLUE` enum case
 
 ## 2.0.0 (2025-03-18)
 * Breaking Changes