+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 
 # PayPal iOS SDK Release Notes
+
+## unreleased
+
+* PayPalWebPayments
+  * Update `PayPalWebCheckoutClient.start()` to recognize `cancel` opType in the deep link urls (fixes #361)
+  * Update `PayPalWebCheckoutClient.vault()` to recognize `cancel` path in deep link urls (fixes #361)
   
 ## 2.0.0 (2025-03-18)
 * Breaking Changes